Corrected a bug
when you place a clip that is shorter than the area (in or out)

+++ __init__.py
@@ -550,14 +550,31 @@ def create_new_log_file():
 
 # Trim an area regarding the IN and OUT points
 def trim_area(scene, inpoint, outpoint):
-	bpy.ops.sequencer.select_all(action = "SELECT")
-	bpy.context.scene.frame_current = inpoint
-	bpy.ops.sequencer.cut(frame=inpoint, type='SOFT', side='LEFT')
-	bpy.ops.sequencer.delete()
-	bpy.ops.sequencer.select_all(action = "SELECT")
-	bpy.context.scene.frame_current = outpoint
-	bpy.ops.sequencer.cut(frame=outpoint, type='SOFT', side='RIGHT')
-	bpy.ops.sequencer.delete()  
+
+	# is it safe to cut on the left?
+	safe = False
+	strips = bpy.context.scene.sequence_editor.sequences
+	for s in strips:
+		if s.type != 'COLOR' and s.frame_final_start < bpy.context.scene.frame_start :
+			safe = True
+			break
+	if safe:
+		bpy.ops.sequencer.select_all(action = "SELECT")
+		bpy.context.scene.frame_current = inpoint
+		bpy.ops.sequencer.cut(frame=inpoint, type='SOFT', side='LEFT')
+		bpy.ops.sequencer.delete()
+	# is it safe to cut on the right?
+	safe = False
+	strips = bpy.context.scene.sequence_editor.sequences
+	for s in strips:
+		if s.type != 'COLOR' and s.frame_final_end > bpy.context.scene.frame_end:
+			safe = True
+			break
+	if safe:
+		bpy.ops.sequencer.select_all(action = "SELECT")
+		bpy.context.scene.frame_current = outpoint
+		bpy.ops.sequencer.cut(frame=outpoint, type='SOFT', side='RIGHT')
+		bpy.ops.sequencer.delete()  
 	bpy.context.scene.frame_current = inpoint
 
 # Initialization ---------------------------------------------------------
